requirements.txt
PyQt6==6.8.0
google-generativeai==0.8.3
python-dotenv==1.0.1